Prod grade WS+Http server with Security is ready.
Prod grade WS+Http server with Security is ready.
Now, there is an authentication done before the user connects to a websocket. And the auth is a prod grade & much thought of.
Also, the WS-instance is accessible through redux in any component.
The WS is connected as user goes to chatroom and is disconnected as user goes back to home page.
Also, the backend schema and auth middleware is setup too.
Now, the rest of the tasks will be done very swiftly.